+++ instrumentation/github.com/Shopify/sarama/otelsarama/consumer_group_test.go
@@ -17,4 +17,4 @@ package otelsarama
 // TODO: add test for consumer group
 // Currently, sarama does not have a mock consumer group, so it's hard to
 // write a unit test.
-// Related PR: https://github.com/Shopify/sarama/pull/1750
+// Related PR: https://github.com/IBM/sarama/pull/1750