+++ instrumentation/github.com/Shopify/sarama/otelsarama/consumer_test.go
@@ -17,8 +17,8 @@ package otelsarama
 import (
 	"testing"
 
-	"github.com/Shopify/sarama"
-	"github.com/Shopify/sarama/mocks"
+	"github.com/IBM/sarama"
+	"github.com/IBM/sarama/mocks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"