+++ instrumentation/github.com/Shopify/sarama/otelsarama/example/producer/producer.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Shopify/sarama"
+	"github.com/IBM/sarama"
 
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/codes"