+++ instrumentation/github.com/Shopify/sarama/otelsarama/producer.go
@@ -21,7 +21,7 @@ import (
 	"strconv"
 	"sync"
 
-	"github.com/Shopify/sarama"
+	"github.com/IBM/sarama"
 
 	"go.opentelemetry.io/otel/codes"