+++ instrumentation/github.com/Shopify/sarama/otelsarama/producer_test.go
@@ -18,8 +18,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Shopify/sarama"
-	"github.com/Shopify/sarama/mocks"
+	"github.com/IBM/sarama"
+	"github.com/IBM/sarama/mocks"
 	"github.com/stretchr/testify/assert"
 
 	oteltrace "go.opentelemetry.io/otel/trace"